Fixed warning and designchanges in Select utility

+++ DjOrm/Utilities/CrudOperations/DeleteUtility.cs
@@ -8,8 +8,8 @@ public class DeleteUtility<T> : Utility, IDeleteUtility<T>
     public async Task DeleteData(T input)
     {
         if (input == null) return;
-        var pk = (await GetNameValueOfProperty(input, new List<Type>() { typeof(PrimaryKeyAttribute) }))?.FirstOrDefault();
-        if (pk is null) return;
+        var pk = (await GetNameValueOfProperty(input, new List<Type>() { typeof(PrimaryKeyAttribute) }))?.First();
+        if (pk is null || pk.Value.Item2 is null) return;
         var command = $"DELETE FROM {GetObjName(input)} WHERE {pk.Value.Item1} = {AppendQuotes(pk.Value.Item2, pk.Value.Item2.GetType())};";
         await _dbConnect.ExecuteCommand(command);
     }
